app/pyproject.toml
[tool.poetry]
name = "adk-kubernetes-agent"
version = "0.1.0"
description = "Weather and Time Agent using Google ADK with clean hexagonal architecture"
authors = ["Alejandro García Osés <alejandro@example.com>"]
maintainers = ["Alejandro Garcia <alejandro@example.com>"]
readme = "README.md"
license = "MIT"
keywords = ["agent", "google-adk", "fastapi", "hexagonal-architecture", "weather"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]

[tool.poetry.urls]
Homepage = "https://github.com/alejandro-garcia/adk-kubernetes"
Repository = "https://github.com/alejandro-garcia/adk-kubernetes.git"
Issues = "https://github.com/alejandro-garcia/adk-kubernetes/issues"
Documentation = "https://github.com/alejandro-garcia/adk-kubernetes/blob/main/README.md"

[tool.poetry.scripts]
adk-agent = "main:start_development_server"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.104.1"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
pydantic = ">=2.5.0"
google-adk = ">=0.1.0"
google-genai = ">=0.1.0"
aiohttp = ">=3.9.0"
python-multipart = ">=0.0.6"
vyper-config = "^1.2.1"
psycopg2-binary = ">=2.9.11"
redis = "^7.1.0"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.1.6"
pytest = ">=7.4.3"
pytest-asyncio = ">=0.21.1"
pytest-cov = ">=4.1.0"
pytest-mock = ">=3.12.0"
httpx = ">=0.25.2"
mypy = ">=1.7.0"
types-aiofiles = ">=23.2.0.0"

[tool.poetry.group.test.dependencies]
pytest = ">=7.4.3"
pytest-asyncio = ">=0.21.1"
pytest-cov = ">=4.1.0"
pytest-mock = ">=3.12.0"
httpx = ">=0.25.2"
litellm = "^1.78.6"

[tool.poetry.group.lint.dependencies]
ruff = ">=0.1.6"
mypy = ">=1.7.0"
types-aiofiles = ">=23.2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Ruff configuration
[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.11+
target-version = "py311"

# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
# Also enable isort (`I`), pydocstyle (`D`), and more
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings (includes W291, W293 for whitespace)
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
]

lint.ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
]

# Allow fix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["D", "S101", "ARG"]  # Allow missing docstrings and asserts in tests
"**/test_*.py" = ["D", "S101", "ARG"]
"**/*_test.py" = ["D", "S101", "ARG"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "contract: Contract tests",
    "slow: Slow tests",
    "expensive: Expensive tests that may incur costs",
    "asyncio: Async tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    ".venv/*",
    "*/site-packages/*",
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "google.*",
    "uvicorn.*",
]
ignore_missing_imports = true